[PWR-267] Unified login screen for admins and users into one singular mask. Moved most login related attributes from AdminState etc to corresponding LoginXYZ files + corresponding refactoring of global constants.
