+++ public/localization/de-DE.json
@@ -85,6 +85,11 @@
   "ProfileEntryType.EDUCATION.AutoCompleteLabel": "Ausbildung",
   "ProfileEntryType.EDUCATION.EditHeader": "Ausbildung Bearbeiten",
   "ProfileEntryType.EDUCATION.Info": "Deine fachliche, schulische oder akademische Ausbildung",
+  "Profile.Login.Remember": "Login merken",
+  "Profile.Login.Rejected": "Ungültiger Nutzername oder Passwort",
+  "Profile.Login.Title": "Login",
+  "Profile.Login.SubTitle": "Weiter zu HBT Power",
+  "Profile.Login.Unavailable": "Login nicht verfügbar",
   "Profile.Project.AddProject": "Projekt Hinzufügen",
   "Profile.Project.NoneSelected": "Kein Projekt Ausgewählt",
   "Profile.Project.List.Title": "{0} für {1}",
@@ -131,9 +136,6 @@
   "ConsultantTile.CreateConsultant": "Neuer Berater",
   "UserLogin.InvalidName": "Ungültiges Kürzel",
   "SkillTree.SearchAndAddSkill": "Skill suchen/hinzufügen",
-  "AdminClient.Login.Remember": "Login merken",
-  "AdminClient.Login.Rejected": "Ungültiger Nutzername oder Passwort",
-  "AdminClient.Login.Unavailable": "Login nicht verfügbar",
   "AdminClient.Menu.Inbox": "Posteingang",
   "AdminClient.Menu.Trashbin": "Papierkorb",
   "AdminClient.Menu.Statistics": "Statistiken",
