+++ src/model/PwrConstants.ts
@@ -10,9 +10,9 @@ export const COOKIE_INITIALS_NAME = 'consultant_initials';
 
 export const COOKIE_INITIALS_EXPIRATION_TIME = 3;
 
-export const COOKIE_ADMIN_USERNAME = 'pwr_admin_username';
+export const COOKIE_USERNAME = 'pwr_username';
 
-export const COOKIE_ADMIN_PASSWORD = 'pwr_admin_password';
+export const COOKIE_PASSWORD = 'pwr_password';
 
 export const COOKIE_ADMIN_EXPIRATION_TIME = {expires: 3}; // 3 Days expiration time
