+++ src/reducers/login/login-action.ts
@@ -1,0 +1,6 @@
+import {LoginStatus} from '../../model/LoginStatus';
+import {AbstractAction} from '../BaseActions';
+
+export interface ChangeLoginStatusAction extends AbstractAction {
+    status: LoginStatus;
+}
